requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
python-multipart==0.0.12
pypdf==4.3.1
python-docx==1.1.2
openai==1.51.0
python-dotenv==1.0.1
